+++ src/core/background_color.hpp
@@ -6,10 +6,10 @@ class BackgroundColor {
 private:
   Color bl, br, tl, tr;
 
-  Color lerp(const Color &A, const Color &B, float t) const;
+  Color lerp(const Color &A, const Color &B, double t) const;
 
 public:
   BackgroundColor(Color bl, Color br, Color tl, Color tr);
   ~BackgroundColor() = default;
-  Color sampleUV(float u, float v) const;
+  Color sampleUV(double u, double v) const;
 };
